implemented waitlist logic

+++ frontend/src/components/DeleteCourse.jsx
@@ -11,11 +11,10 @@ const DeleteCourse = ({ updateRender, course, isSniperRunning }) => {
 
     useEffect(() => {
         const handleKeyDown = async (e) => {
-            if (e.key === "Enter" && isConfirmModalOpen) {
+            if (e.key === "Enter" && isConfirmModalOpen && !disabled) {
                 deleteCourse();
             }
             if (e.key === "Escape" && isConfirmModalOpen) {
-                setDisabled(false);
                 setIsConfirmModalOpen(false);
             }
         };
@@ -26,6 +25,7 @@ const DeleteCourse = ({ updateRender, course, isSniperRunning }) => {
     }, [isConfirmModalOpen]);
 
     const deleteCourse = async () => {
+        setDisabled(true);
         const uid = currentUser.uid;
 
         try {
@@ -43,12 +43,13 @@ const DeleteCourse = ({ updateRender, course, isSniperRunning }) => {
         }
 
         setIsConfirmModalOpen(false);
+        setDisabled(true);
     };
 
     return (
         <div className="relative font-sans font-medium text-center align-middle transition-all select-none place-content-center ">
             <button
-                className={`disabled:opacity-50 disabled:shadow-none disabled:pointer-events-none w-10 max-w-[40px] h-10 max-h-[40px] rounded-lg text-xs ${
+                className={`disabled:opacity-50 disabled:shadow-none disabled:pointer-events-none rounded-lg text-xs ${
                     isSniperRunning ? "hover:cursor-not-allowed" : "hover:text-red-500"
                 }`}
                 onClick={(e) => {
@@ -56,19 +57,30 @@ const DeleteCourse = ({ updateRender, course, isSniperRunning }) => {
                     if (isSniperRunning) {
                         toast.error("Cannot delete courses while sniping.");
                     } else {
-                        setDisabled(true);
                         setIsConfirmModalOpen(true);
                     }
                 }}
                 disabled={disabled}
             >
-                <Trash2 />
+                <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke-width="1.5"
+                    stroke="currentColor"
+                    className="items-center justify-center size-6"
+                >
+                    <path
+                        stroke-linecap="round"
+                        stroke-linejoin="round"
+                        d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0"
+                    />
+                </svg>
             </button>
             <div>
                 <ConfirmModal
                     isOpen={isConfirmModalOpen}
                     onClose={() => {
-                        setDisabled(false);
                         setIsConfirmModalOpen(false);
                     }}
                     onConfirm={async () => {
